TAP_final/requirements.txt
# to use this file, do: pip install -r requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
bert-score==0.3.13
Brotli==1.0.9
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
datasets==3.0.2
dill==0.3.8
einops=0.8.1
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
h2==4.1.0
hpack==4.0.0
huggingface_hub==0.26.5
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
nltk=3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.0.3
peft==0.13.2
pillow==10.4.0
pip==24.3.1
psutil==7.0.0
pyarrow==17.0.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sacremoses==0.1.1
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.9.4
zstandard==0.19.0
